Fix a security issue with the scripts
A script could be put into any arbitrary place on the system, meaning a
remote user could add script=xxx to the url and run any arbitrary
program on the system.

Pass in a script directory with the URL, put the specified script on
that and make sure it stays in that directory.

Signed-off-by: Corey Minyard <minyard@acm.org>